show telemetry popup before onboarding flow, enable sentry only after user consent
